+++ src/main/java/com/project/humanresource/controller/AssignmentController.java
@@ -2,7 +2,7 @@ package com.project.humanresource.controller;
 
 import com.project.humanresource.entity.Assignment;
 import com.project.humanresource.service.AssignmentService;
-import com.project.humanresource.utility.BaseResponse;
+import com.project.humanresource.dto.response.BaseResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
