+++ src/main/java/com/project/humanresource/dto/response/BaseResponse.java
@@ -1,14 +1,39 @@
 package com.project.humanresource.dto.response;
 
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
 public class BaseResponse<T> {
-    String message;
-    Integer code;
-    T data;
-}
+    private boolean success;
+    private String message;
+    private T data;
+
+    public BaseResponse() {}
+
+    public BaseResponse(boolean success, String message, T data) {
+        this.success = success;
+        this.message = message;
+        this.data = data;
+    }
+
+    public boolean isSuccess() {
+        return success;
+    }
+
+    public void setSuccess(boolean success) {
+        this.success = success;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public T getData() {
+        return data;
+    }
+
+    public void setData(T data) {
+        this.data = data;
+    }
+}
