--- src/main/java/com/project/humanresource/utility/BaseResponse.java
@@ -1,39 +1,0 @@
-package com.project.humanresource.utility;
-
-public class BaseResponse<T> {
-    private boolean success;
-    private String message;
-    private T data;
-
-    public BaseResponse() {}
-
-    public BaseResponse(boolean success, String message, T data) {
-        this.success = success;
-        this.message = message;
-        this.data = data;
-    }
-
-    public boolean isSuccess() {
-        return success;
-    }
-
-    public void setSuccess(boolean success) {
-        this.success = success;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public T getData() {
-        return data;
-    }
-
-    public void setData(T data) {
-        this.data = data;
-    }
-}
